+++ src/main/java/com/example/newProject/Repositories/HouseOwnerRepository.java
@@ -2,7 +2,7 @@ package com.example.newProject.Repositories;
 import com.example.newProject.Entities.HouseOwner;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface HouseOwnerRepository extends JpaRepository<HouseOwner,Integer> {
+public interface HouseOwnerRepository extends JpaRepository<HouseOwner,Long> {
