+++ src/main/java/com/example/newProject/Services/CustomerService.java
@@ -27,7 +27,11 @@ public class CustomerService {
         this.customerRepository = customerRepository;
     }
 
-    public List<Customer> getAllCustomersOfOneHouse(int id){
+
+    //1 tane house çek ve onun içinde olduğu tüm ilişkilerini bul
+    //sonra bu ilişkilerin her birisinin customerını bul ve listeye ekle
+    //sonuçta bir evin tüm customerlarını bulmuş oluruz
+    public List<Customer> getAllCustomersOfOneHouse(Long id){
         House house = houseRepository.findById(id).orElse(null);
         List<HouseCustomer> relaions=havingRelationRepository.findHavingRelationsByHouse(house);
         //fe found all relations of one house
@@ -40,7 +44,7 @@ public class CustomerService {
         return allCustomersOfOneHouse;
     }
 
-
+    //get all customers in db
     public List<Customer> getAllCustomers(){
         return customerRepository.findAll();
     }
