+++ src/main/java/com/example/newProject/api/controllers/HavingRelationController.java
@@ -22,7 +22,7 @@ public class HavingRelationController {
     }
 
     @GetMapping("/getRalationsOfOneCustomer/{id}")
-    public List<HouseCustomer> getRalationsOfOneCustomer(@PathVariable int id){
+    public List<HouseCustomer> getRalationsOfOneCustomer(@PathVariable Long id){
         return houseService.getRalationsOfOneCustomer(id);
     }
 }
